requirements.txt
django==1.9.2
django-leaflet
django-tastypie
django-jquery
django-reversion
django-extensions
psycopg2
django-geojson
django-mapit
django-suit
django-grappelli-filters
notebook
ipython
autopep8
flake8
yapf
